+++ auto_route/example/lib/mobile/screens/books/book_list_page.dart
@@ -11,6 +11,7 @@ class BookListPage extends StatefulWidget {
 class _BookListPageState extends State<BookListPage> {
   @override
   Widget build(BuildContext context) {
+   
     var booksDb = BooksDBProvider.of(context);
     return Scaffold(
       body: ListView(