+++ auto_route/lib/src/router/controller/auto_route_guard.dart
@@ -66,6 +66,7 @@ abstract class AutoRedirectGuardBase extends AutoRouteGuard
 
 abstract class AutoRedirectGuard extends AutoRedirectGuardBase {
   NavigationResolver? _redirectResolver;
+
   @protected
   void redirect(PageRouteInfo route,
       {required NavigationResolver resolver}) async {
@@ -77,9 +78,11 @@ abstract class AutoRedirectGuard extends AutoRedirectGuardBase {
     router.push(route).then((_) => _redirectResolver = null);
     await resolver._completer.future;
     if (router.current.name == route.routeName) {
-      router._removeLast(notify: false);
       router.markUrlStateForReplace();
     }
+
+    router.removeWhere((r) => r.name == route.routeName, notify: false);
+
     _redirectResolver = null;
   }
 
@@ -107,7 +110,7 @@ abstract class ReevaluationStrategy {
       RePushFirstGuardedAndUp;
 
   const factory ReevaluationStrategy.removeFirstGuardedRouteAndUp() =
-      _removeFirstGuardedAndUp;
+      _RemoveFirstGuardedAndUp;
 }
 
 class RePushAllStrategy extends ReevaluationStrategy {
@@ -122,7 +125,16 @@ class RePushAllStrategy extends ReevaluationStrategy {
       router._removeRoute(route, notify: false);
     }
 
-    router._pushAllGuarded(List.of(stackData.map((e) => e.route)));
+    final routesToPush = <RouteMatch>[];
+    for(final existingMatch in stackData.map((e) => e.route)){
+      final routeToPush = router.matcher.matchByRoute(
+        existingMatch.toPageRouteInfo(),
+      );
+      if(routeToPush!= null) {
+        routesToPush.add(routeToPush);
+      }
+    }
+    router._pushAllGuarded(routesToPush);
   }
 }
 
@@ -141,7 +153,13 @@ class RePushFirstGuarded extends ReevaluationStrategy {
     for (final route in routesToRemove) {
       router._removeRoute(route, notify: false);
     }
-    router._pushAllGuarded([routes[firstGuardedRouteIndex]]);
+    // resolve initial child routes if there are any
+    final routeToPush = router.matcher.matchByRoute(
+      routes[firstGuardedRouteIndex].toPageRouteInfo(),
+    );
+    if(routeToPush!= null) {
+      router._pushAllGuarded([routeToPush]);
+    }
   }
 }
 
@@ -159,15 +177,26 @@ class RePushFirstGuardedAndUp extends ReevaluationStrategy {
     for (final route in routesToRemove) {
       router._removeRoute(route, notify: false);
     }
-    router._pushAllGuarded(routes.sublist(
+
+
+    final routesToPush = <RouteMatch>[];
+    for(final existingMatch in routes.sublist(
       firstGuardedRouteIndex,
       routes.length,
-    ));
+    )){
+      final routeToPush = router.matcher.matchByRoute(
+        existingMatch.toPageRouteInfo(),
+      );
+      if(routeToPush!= null) {
+       routesToPush.add(routeToPush);
+      }
+    }
+    router._pushAllGuarded(routesToPush);
   }
 }
 
-class _removeFirstGuardedAndUp extends ReevaluationStrategy {
-  const _removeFirstGuardedAndUp() : super._();
+class _RemoveFirstGuardedAndUp extends ReevaluationStrategy {
+  const _RemoveFirstGuardedAndUp() : super._();
 
   @override
   void reevaluate(AutoRedirectGuardBase guard, StackRouter router) {